moved Tagging to vsphere top level and added TaggerFactory
