+++ vsphere/src/main/java/com/elastisys/scale/cloudpool/vsphere/tagger/TaggerFactory.java
@@ -1,0 +1,12 @@
+package com.elastisys.scale.cloudpool.vsphere.tagger;
+
+import com.elastisys.scale.cloudpool.vsphere.tagger.impl.CustomAttributeTagger;
+
+public class TaggerFactory {
+
+    private TaggerFactory() {}
+
+    public static Tagger getTagger(){
+        return new CustomAttributeTagger();
+    }
+}
